backend/requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
PyJWT==2.9.0
Flask-CORS==4.0.1
PyMySQL==1.1.1
python-dotenv==1.0.1
Pillow==10.3.0
cryptography==44.0.2
